Refactor exceptions
move auth specific exceptions into exceptions folder inside auth.
