+++ pcs-config/app/com/microsoft/azure/iotsolutions/uiconfig/webservice/auth/OpenIdConnectJwtValidation.java
@@ -4,7 +4,7 @@ package com.microsoft.azure.iotsolutions.uiconfig.webservice.auth;
 
 import com.google.inject.Inject;
 import com.google.inject.Singleton;
-import com.microsoft.azure.iotsolutions.uiconfig.services.exceptions.NotAuthorizedException;
+import com.microsoft.azure.iotsolutions.uiconfig.webservice.auth.exceptions.*;
 import com.nimbusds.jose.*;
 import com.nimbusds.jose.jwk.source.JWKSource;
 import com.nimbusds.jose.jwk.source.RemoteJWKSet;
